Add CMake option to disable/enable TTS support
